Inserindo novos tratamentos nos servicos REST

+++ aes/src/java/aes/service/AgendaAppointmentFacadeREST.java
@@ -61,7 +61,7 @@ public class AgendaAppointmentFacadeREST extends AbstractFacade<AgendaAppointmen
         try {
             appointmentDao.insert(appointment, em);
             return Response.status(Response.Status.CREATED).build();
-        } catch (SQLException ex) {
+        } catch (SQLException | RuntimeException ex) {
             Logger.getLogger(AgendaAppointmentFacadeREST.class.getName()).log(Level.SEVERE, null, ex);
             return Response.status(Response.Status.BAD_REQUEST).build();
         }
@@ -74,7 +74,7 @@ public class AgendaAppointmentFacadeREST extends AbstractFacade<AgendaAppointmen
         try {
             appointmentDao.update(appointment, em);
             return Response.status(Response.Status.OK).build();
-        } catch (SQLException ex) {
+        } catch (SQLException | RuntimeException ex) {
             Logger.getLogger(AgendaAppointmentFacadeREST.class.getName()).log(Level.SEVERE, null, ex);
             return Response.status(Response.Status.BAD_REQUEST).build();
         }
@@ -83,11 +83,15 @@ public class AgendaAppointmentFacadeREST extends AbstractFacade<AgendaAppointmen
     @DELETE
     @Path("delete/{id}")
     public Response delete(@PathParam("id") Long id) {
+        AgendaAppointment app;
         try {
-            AgendaAppointment app = new AgendaAppointment(id);
-            appointmentDao.delete(app, em);
-            return Response.status(Response.Status.OK).build();
-        } catch (SQLException ex) {
+            if(appointmentDao.find(id, em) != null) {
+                app = new AgendaAppointment(id);
+                appointmentDao.delete(app, em);
+                return Response.status(Response.Status.OK).build();
+            }
+            return Response.status(Response.Status.NOT_FOUND).build();
+        } catch (SQLException | RuntimeException ex) {
             Logger.getLogger(AgendaAppointmentFacadeREST.class.getName()).log(Level.SEVERE, null, ex);
             return Response.status(Response.Status.BAD_REQUEST).build();
         }
@@ -106,7 +110,7 @@ public class AgendaAppointmentFacadeREST extends AbstractFacade<AgendaAppointmen
     public Response findAllAppointments() {
         try {
             return Response.ok().entity(appointmentDao.list(em)).build();
-        } catch (SQLException ex) {
+        } catch (SQLException | RuntimeException ex) {
             Logger.getLogger(AgendaAppointmentFacadeREST.class.getName()).log(Level.SEVERE, null, ex);
             return Response.status(Response.Status.BAD_REQUEST).build();
         }
@@ -118,7 +122,7 @@ public class AgendaAppointmentFacadeREST extends AbstractFacade<AgendaAppointmen
     public Response findAllCurrentByUser(@PathParam("userId") Long userId) {
         try {
             return Response.ok().entity(appointmentDao.listCurrentByUser(userId, em)).build();
-        } catch (SQLException ex) {
+        } catch (SQLException | RuntimeException ex) {
             Logger.getLogger(AgendaAppointmentFacadeREST.class.getName()).log(Level.SEVERE, null, ex);
             return Response.status(Response.Status.BAD_REQUEST).build();
         }
@@ -130,7 +134,7 @@ public class AgendaAppointmentFacadeREST extends AbstractFacade<AgendaAppointmen
     public Response findAllByConsultant(@PathParam("consultantId") Long consultantId) {
         try {
             return Response.ok().entity(appointmentDao.list("consultant.id", consultantId, em)).build();
-        } catch (SQLException ex) {
+        } catch (SQLException | RuntimeException ex) {
             Logger.getLogger(AgendaAppointmentFacadeREST.class.getName()).log(Level.SEVERE, null, ex);
             return Response.status(Response.Status.BAD_REQUEST).build();
         }
